requirements.txt
# 기존 의존성
fastapi==0.95.0
uvicorn==0.22.0
python-dotenv==1.0.0
python-jose==3.3.0
sqlalchemy==2.0.9
psycopg2-binary==2.9.6
alembic>=1.9.0
pydantic==1.10.7
pydantic-settings==0.2.5

# Redis 캐싱 
redis==4.5.4

# 추가된 의존성
cryptography==40.0.1
passlib==1.7.4
python-multipart==0.0.6
bcrypt==4.0.1
aiohttp==3.8.4
websockets==11.0.2
pytest==7.3.1
requests==2.28.2
more-itertools==9.1.0
python-dateutil==2.8.2

# 보안 강화 의존성
pycryptodome==3.19.0  # AES-256 암호화를 위한 패키지
pyjwt==2.8.0  # JWT 인증 개선
oauthlib==3.2.2  # OAuth 인증을 위한 패키지
httpx==0.24.1  # HTTP 클라이언트 (HTTPS 지원)
